both schedulers use the same task
